requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
google-api-python-client>=2.0.0
textblob>=0.15.3
python-dotenv>=0.19.0  # optional if using .env